feat(debug): scope-precise breakpoints
Continuing work done in #5957, these changes add column precision
to breakpoints for those set using the keymap command. For those set
with the mouse, preserve the current behaviour, concretely, add a line
breakpoint. This allows for more fine grained control, allowing, for
debuggers which support it, to add breakpoints to scopes such as lambda
functions, callbacks and more.

Change how the stack frame line is highlighted, using a full line for
breakpoints which apply to a line and highlight only the scope for those
that operate only on those scopes.

Closes: #6238
Signed-off-by: Filip Dutescu <filip.dutescu@gmail.com>